Rewriting KFactors.
- Added HyperplaneRanker to rank features by distance from points.
- Added IndependentFactorAssignment to calculate point memberships in features.
- Added new test file to validate behavior in 1d.

KFactors is still broken. Tests don't pass.

I'm having to rewrite KFactors almost from scratch.

So much for trust AI to write huge swaths of code.

I didn't touch this for like two months after Claude Opus wrote it all. I knew it was going to be pain. But it was a beautiful piece of work until I looked closer.